Project/Backend/pyproject.toml
[project]
name = "project"
version = "0.1.0"
description = ""
authors = [
    {name = "Prasanna-Kudale",email = "kudaleprasanna36@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "portia-sdk-python[all] (>=0.7.2,<0.8.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "python-telegram-bot (>=22.3,<23.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "beautifulsoup4 (>=4.13.4,<5.0.0)",
    "fastapi (>=0.116.1,<0.117.0)",
    "uvicorn (>=0.35.0,<0.36.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "pydantic (>=2.11.7,<3.0.0)"
]

[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
